Recreate requirements.txt from freeze
Versions are good to include, and this will make it easier to update in the future.

I left in pytz and psycopg2-binary even though we do not use them, since I assume they were put there so someone else can use them (or because we'll need them). If we don't, I'll just remove them later.