Taxonomic Data Processing Script
My best attempt at parsing the Catalogue of Life and turning it into something we can use for ourselves. It was horribly slow trying to use the whole thing, so I ended up needing to have it just take smaller branches. Not perfect, but about as good as it is going to get.