Adds departmental Note permissions
For each of "ceo", "hr", "board", and "volunteer", there are "_add_note", "_view_note", and "_delete_note" permissions. If the new field on the Note model is empty, the note is considered accessible to anyone with standard note permissions. Otherwise, you can only see a note if you have any of the required view permissions. For creation and deletion, you need ALL of the relevant permissions.

Migrates the database.